move test to its module
