+++ skore/tests/unit/sklearn/train_test_split/test_train_test_split.py
@@ -10,19 +10,6 @@ from skore.sklearn.train_test_split.warning import (
 )
 
 
-def test_check_high_class_imbalance_too_few_examples():
-    y_test = [0] * 100
-    y_labels = [0, 1]
-    check_passed = HighClassImbalanceTooFewExamplesWarning.check(
-        y_test=y_test,
-        y_labels=y_labels,
-        stratify=None,
-        ml_task="multiclass-classification",
-    )
-
-    assert not check_passed
-
-
 def test_train_test_split_warns():
     with pytest.warns(HighClassImbalanceWarning, match=HighClassImbalanceWarning.MSG):
         train_test_split([[1]] * 4, [0, 1, 1, 1])
